refactor: classify the custom block node handler
