+++ libs/toastmark/src/commonmark/blockHandlers.ts
@@ -9,7 +9,8 @@ import {
   tableDelimRow,
   tableDelimCell
 } from './gfm/tableBlockHandler';
-import { ListNode, BlockNode, CodeBlockNode, HtmlBlockNode, CustomBlockNode } from './node';
+import { customBlock } from './custom/customBlockHandler';
+import { ListNode, BlockNode, CodeBlockNode, HtmlBlockNode } from './node';
 import {
   peek,
   isBlank,
@@ -18,8 +19,7 @@ import {
   reClosingCodeFence,
   CODE_INDENT,
   C_OPEN_BRACKET,
-  C_GREATERTHAN,
-  reClosingCustomBlock
+  C_GREATERTHAN
 } from './blockHelper';
 import { unescapeString } from './common';
 
@@ -262,40 +262,6 @@ const refDef: BlockHandler = {
     return Process.Stop;
   },
   finalize() {},
-  canContain() {
-    return false;
-  },
-  acceptsLines: true
-};
-
-const customBlock: BlockHandler = {
-  continue(parser, container: CustomBlockNode) {
-    const ln = parser.currentLine;
-    const match = ln.slice(parser.nextNonspace).match(reClosingCustomBlock);
-    if (match) {
-      // closing custom block - we're at end of line, so we can return
-      parser.lastLineLength = match[0].length;
-      parser.finalize(container as BlockNode, parser.lineNumber);
-      return Process.Finished;
-    }
-    return Process.Go;
-  },
-  finalize(_, block: CustomBlockNode) {
-    if (block.stringContent === null) {
-      return;
-    }
-    if (block.isCustomBlock) {
-      // first line becomes info string
-      const content = block.stringContent;
-      const newlinePos = content.indexOf('\n');
-      const firstLine = content.slice(0, newlinePos);
-      const rest = content.slice(newlinePos + 1);
-      const infoString = firstLine.match(/^(\s*)(.*)/);
-      block.info = unescapeString(infoString![2].trim());
-      block.literal = rest;
-    }
-    block.stringContent = null; // allow GC
-  },
   canContain() {
     return false;
   },
